requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
Django==5.2.2
djangorestframework==3.16.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
ruff==0.11.13
sqlparse==0.5.3
uritemplate==4.2.0
